Update README.md
Basically created an interactive display that shows the bacteria present in an individuals belly button. It allows you to select a Sample ID and brings up the biographical metadata of the person and shows a bar graph of the top 10 bacteria and a bubble chart showing the size of the bacterias.